dl_inference/requirements.txt
azureml-sdk==1.0.21
azure-common==1.1.26
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==11.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.3.0
docker==4.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
gast==0.2.2
requests==2.25.0
requests-oauthlib==1.3.0
PyYaml